+++ src/main/java/nextstep/courses/domain/course/Course.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.course;
 
 import nextstep.courses.domain.session.Session;
 import nextstep.courses.domain.session.Sessions;
