+++ src/main/java/nextstep/courses/domain/enrollment/EnrollmentCandidate.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.session;
+package nextstep.courses.domain.enrollment;
 
 public class EnrollmentCandidate {
     private final Long sessionId;
