+++ src/main/java/nextstep/courses/domain/image/SessionImages.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.session.image;
+package nextstep.courses.domain.image;
 
 import java.util.ArrayList;
 import java.util.Collections;
