+++ src/main/java/nextstep/courses/domain/session/SessionInfo.java
@@ -1,7 +1,7 @@
 package nextstep.courses.domain.session;
 
-import nextstep.courses.domain.session.image.SessionImage;
-import nextstep.courses.domain.session.image.SessionImages;
+import nextstep.courses.domain.image.SessionImage;
+import nextstep.courses.domain.image.SessionImages;
 
 import java.time.LocalDate;
