+++ src/main/java/nextstep/courses/infrastructure/JdbcSessionRepository.java
@@ -9,8 +9,8 @@ import nextstep.courses.domain.session.SessionInfo;
 import nextstep.courses.domain.session.SessionRepository;
 import nextstep.courses.domain.session.SessionType;
 import nextstep.courses.domain.session.Sessions;
-import nextstep.courses.domain.session.image.SessionImage;
-import nextstep.courses.domain.session.image.SessionImages;
+import nextstep.courses.domain.image.SessionImage;
+import nextstep.courses.domain.image.SessionImages;
 import org.springframework.jdbc.core.JdbcOperations;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
