+++ src/test/java/nextstep/courses/domain/enrollment/EnrollmentTest.java
@@ -1,5 +1,9 @@
-package nextstep.courses.domain.session;
+package nextstep.courses.domain.enrollment;
 
+import nextstep.courses.domain.session.FreeSessionType;
+import nextstep.courses.domain.session.PaidSessionType;
+import nextstep.courses.domain.session.RecruitmentStatus;
+import nextstep.courses.domain.session.SessionType;
 import nextstep.payments.domain.Payment;
 import org.junit.jupiter.api.Test;
