+++ src/test/java/nextstep/courses/domain/image/ImageDimensionTest.java
@@ -1,6 +1,5 @@
-package nextstep.courses.domain.session.image;
+package nextstep.courses.domain.image;
 
-import nextstep.courses.domain.session.image.ImageDimension;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
