+++ src/test/java/nextstep/courses/domain/session/SessionsTest.java
@@ -1,6 +1,6 @@
 package nextstep.courses.domain.session;
 
-import nextstep.courses.domain.session.image.SessionImage;
+import nextstep.courses.domain.image.SessionImage;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
